Moved min and max
